// ev-backoffice/client/src/app/dashboard/articles/models/article.model.ts
export class Article {
  id: number;
  author: string;
  location: string;
  locationId: string;
  locationName: string;
  title: string;
  content: string;
  views: number;
  words: number;
  dateSubmitted: string;
  dateDispositioned: string;
  approved: boolean;
  url?: string;
  organizationName: string;
}